src/third_party/blink/web_tests/external/wpt/webrtc/receiver-track-live.https-expected.txt

This is a testharness.js-based test.
PASS Setup audio call
FAIL Inactivate the audio transceiver assert_equals: expected "live" but got "ended"
FAIL Reactivate the audio transceiver assert_equals: expected "live" but got "ended"
PASS Clean-up
Harness: the test ran to completion.
